src/requirements.txt
babel
click
humanize
Jinja2
pytz
PyYAML
python-jose
PyGithub
lpshipit>=0.4
joblib
GitPython
tox
launchpadlib